pyproject.toml
[tool.poetry]
name = "factorio-manager"
version = "0.1.0"
description = ""
authors = ["asoul-rec <asoulranbi@gmail.com>"]
readme = "README.md"
packages = [
    {include = "facmgr"}
]

# [tool.poetry.scripts]
# factorio-manager-server = 'server:__name__'

[tool.poetry.dependencies]
python = "^3.13"
grpcio = "^1.68.0"
grpcio-tools = "^1.68.0"
pyrofork = "^2.3.50"
TgCrypto = "^1.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
